Cleanup dependency class organization
- Move classes into their own files
- Remove classes that aren't used
