+++ lib/inspec/requirement.rb
@@ -1,12 +1,16 @@
 # encoding: utf-8
+require 'inspec/fetcher'
+
 module Inspec
   class Requirement
     attr_reader :name, :dep, :cwd, :opts
+
     def initialize(name, version_constraints, cwd, opts)
       @name = name
       @dep = Gem::Dependency.new(name,
                                  Gem::Requirement.new(Array(version_constraints)),
                                  :runtime)
+      @fetcher = fetcher_for_options(opts)
       @opts = opts
       @cwd = cwd
     end
@@ -14,6 +18,19 @@ module Inspec
     def matches_spec?(spec)
       params = spec.profile.metadata.params
       @dep.match?(params[:name], params[:version])
+    end
+
+    def fetcher_for_options(opts)
+      f = if opts[:path]
+            Fetchers::Local.resolve(File.expand_path(opts[:path], @cwd))
+          elsif opts[:url]
+            Fetchers::Url.resolve(opts[:url])
+          else
+            fail "No known fetcher for dependency #{name} (options: #{opts})"
+          end
+
+      fail "Unable to resolve source for #{name} (options: #{opts})" if f.nil?
+      f
     end
 
     def pull
